+++ lutra/lutra/src/pull_schema.rs
@@ -3,7 +3,7 @@ use arrow::datatypes::{DataType, SchemaRef};
 use connector_arrow::api::SchemaGet;
 use itertools::Itertools;
 use prqlc::ast::{PrimitiveSet, Stmt, Ty, TyKind, TyTupleField, VarDef};
-use prqlc::{Error, WithErrorInfo};
+use prqlc::{parser_error::Error, WithErrorInfo};
 
 use crate::ProjectCompiled;
