--- prqlc/prqlc-parser/src/err/mod.rs
@@ -1,3 +1,0 @@
-pub mod error;
-pub mod parse_error;
-mod test;
