+++ prqlc/prqlc-parser/src/lexer/mod.rs
@@ -3,11 +3,56 @@ use chumsky::{
     prelude::*,
     text::{newline, Character},
 };
-use prqlc_ast::token::{Literal, Token, TokenKind, ValueAndUnit};
 use serde::{Deserialize, Serialize};
 
+use crate::error::{Error, ErrorSource, Reason, WithErrorInfo};
+use crate::lexer;
+use crate::lexer::lr::{Literal, Token, TokenKind, ValueAndUnit};
+use crate::span::Span;
+
+pub mod lr;
+mod test_lexer;
+
+pub fn lex_string_recovery(source: &str, source_id: u16) -> (Option<Vec<Token>>, Vec<Error>) {
+    let (tokens, lex_errors) = ::chumsky::Parser::parse_recovery(&lexer::lexer(), source);
+
+    let errors = lex_errors
+        .into_iter()
+        .map(|e| convert_lexer_error(source, e, source_id))
+        .collect();
+
+    (tokens, errors)
+}
+
+pub fn lex_source(source: &str) -> Result<TokenVec, Vec<Error>> {
+    lexer().parse(source).map(TokenVec).map_err(|e| {
+        e.into_iter()
+            .map(|x| convert_lexer_error(source, x, 0))
+            .collect()
+    })
+}
+
+fn convert_lexer_error(source: &str, e: chumsky::error::Cheap<char>, source_id: u16) -> Error {
+    // We want to slice based on the chars, not the bytes, so can't just index
+    // into the str.
+    let found = source
+        .chars()
+        .skip(e.span().start)
+        .take(e.span().end() - e.span().start)
+        .collect();
+    let span = Some(Span {
+        start: e.span().start,
+        end: e.span().end,
+        source_id,
+    });
+
+    Error::new(Reason::Unexpected { found })
+        .with_span(span)
+        .with_source(ErrorSource::Lexer(e))
+}
+
 /// Lex chars to tokens until the end of the input
-pub fn lexer() -> impl Parser<char, Vec<Token>, Error = Cheap<char>> {
+fn lexer() -> impl Parser<char, Vec<Token>, Error = Cheap<char>> {
     lex_token()
         .repeated()
         .then_ignore(ignored())
@@ -15,7 +60,7 @@ pub fn lexer() -> impl Parser<char, Vec<Token>, Error = Cheap<char>> {
 }
 
 /// Lex chars to a single token
-pub fn lex_token() -> impl Parser<char, Token, Error = Cheap<char>> {
+fn lex_token() -> impl Parser<char, Token, Error = Cheap<char>> {
     let control_multi = choice((
         just("->").to(TokenKind::ArrowThin),
         just("=>").to(TokenKind::ArrowFat),
@@ -428,195 +473,3 @@ fn end_expr() -> impl Parser<char, (), Error = Cheap<char>> {
 
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
 pub struct TokenVec(pub Vec<Token>);
-
-#[cfg(test)]
-mod test {
-    use insta::assert_debug_snapshot;
-    use insta::assert_snapshot;
-
-    use super::*;
-
-    #[test]
-    fn line_wrap() {
-        assert_debug_snapshot!(TokenVec(lexer().parse(r"5 +
-    \ 3 "
-        ).unwrap()), @r###"
-        TokenVec(
-            [
-                0..1: Literal(Integer(5)),
-                2..3: Control('+'),
-                3..9: LineWrap([]),
-                10..11: Literal(Integer(3)),
-            ],
-        )
-        "###);
-
-        // Comments are included; no newline after the comments
-        assert_debug_snapshot!(TokenVec(lexer().parse(r"5 +
-# comment
-   # comment with whitespace
-  \ 3 "
-        ).unwrap()), @r###"
-        TokenVec(
-            [
-                0..1: Literal(Integer(5)),
-                2..3: Control('+'),
-                3..46: LineWrap([Comment(" comment"), Comment(" comment with whitespace")]),
-                47..48: Literal(Integer(3)),
-            ],
-        )
-        "###);
-
-        // Check display, for the test coverage (use `assert_eq` because the
-        // line-break doesn't work well with snapshots)
-        assert_eq!(
-            format!(
-                "{}",
-                TokenKind::LineWrap(vec![TokenKind::Comment(" a comment".to_string())])
-            ),
-            r#"
-\ # a comment
-"#
-        );
-    }
-
-    #[test]
-    fn numbers() {
-        // Binary notation
-        assert_eq!(
-            literal().parse("0b1111000011110000").unwrap(),
-            Literal::Integer(61680)
-        );
-        assert_eq!(
-            literal().parse("0b_1111000011110000").unwrap(),
-            Literal::Integer(61680)
-        );
-
-        // Hexadecimal notation
-        assert_eq!(literal().parse("0xff").unwrap(), Literal::Integer(255));
-        assert_eq!(
-            literal().parse("0x_deadbeef").unwrap(),
-            Literal::Integer(3735928559)
-        );
-
-        // Octal notation
-        assert_eq!(literal().parse("0o777").unwrap(), Literal::Integer(511));
-    }
-
-    #[test]
-    fn debug_display() {
-        assert_debug_snapshot!(TokenVec(lexer().parse("5 + 3").unwrap()), @r###"
-        TokenVec(
-            [
-                0..1: Literal(Integer(5)),
-                2..3: Control('+'),
-                4..5: Literal(Integer(3)),
-            ],
-        )
-        "###);
-    }
-
-    #[test]
-    fn comment() {
-        assert_debug_snapshot!(TokenVec(lexer().parse("# comment\n# second line").unwrap()), @r###"
-        TokenVec(
-            [
-                0..9: Comment(" comment"),
-                9..10: NewLine,
-                10..23: Comment(" second line"),
-            ],
-        )
-        "###);
-
-        assert_snapshot!(TokenKind::Comment(" This is a single-line comment".to_string()), @r###"
-        # This is a single-line comment
-        "###);
-    }
-
-    #[test]
-    fn doc_comment() {
-        assert_debug_snapshot!(TokenVec(lexer().parse("#! docs").unwrap()), @r###"
-        TokenVec(
-            [
-                0..7: DocComment(" docs"),
-            ],
-        )
-        "###);
-    }
-
-    #[test]
-    fn quotes() {
-        // All these are valid & equal.
-        assert_snapshot!(quoted_string(false).parse(r#"'aoeu'"#).unwrap(), @"aoeu");
-        assert_snapshot!(quoted_string(false).parse(r#"'''aoeu'''"#).unwrap(), @"aoeu");
-        assert_snapshot!(quoted_string(false).parse(r#"'''''aoeu'''''"#).unwrap(), @"aoeu");
-        assert_snapshot!(quoted_string(false).parse(r#"'''''''aoeu'''''''"#).unwrap(), @"aoeu");
-
-        // An even number is interpreted as a closed string (and the remainder is unparsed)
-        assert_snapshot!(quoted_string(false).parse(r#"''aoeu''"#).unwrap(), @"");
-
-        // When not escaping, we take the inner string between the three quotes
-        assert_snapshot!(quoted_string(false).parse(r#""""\"hello\""""#).unwrap(), @r###"\"hello\"###);
-
-        assert_snapshot!(quoted_string(true).parse(r#""""\"hello\"""""#).unwrap(), @r###""hello""###);
-
-        // Escape each inner quote depending on the outer quote
-        assert_snapshot!(quoted_string(true).parse(r#""\"hello\"""#).unwrap(), @r###""hello""###);
-        assert_snapshot!(quoted_string(true).parse(r"'\'hello\''").unwrap(), @"'hello'");
-
-        assert_snapshot!(quoted_string(true).parse(r#"''"#).unwrap(), @"");
-
-        // An empty input should fail
-        quoted_string(false).parse(r#""#).unwrap_err();
-
-        // An even number of quotes is an empty string
-        assert_snapshot!(quoted_string(true).parse(r#"''''''"#).unwrap(), @"");
-
-        // Hex escape
-        assert_snapshot!(quoted_string(true).parse(r"'\x61\x62\x63'").unwrap(), @"abc");
-
-        // Unicode escape
-        assert_snapshot!(quoted_string(true).parse(r"'\u{01f422}'").unwrap(), @"🐢");
-    }
-
-    #[test]
-    fn range() {
-        assert_debug_snapshot!(TokenVec(lexer().parse("1..2").unwrap()), @r###"
-        TokenVec(
-            [
-                0..1: Literal(Integer(1)),
-                1..3: Range { bind_left: true, bind_right: true },
-                3..4: Literal(Integer(2)),
-            ],
-        )
-        "###);
-
-        assert_debug_snapshot!(TokenVec(lexer().parse("..2").unwrap()), @r###"
-        TokenVec(
-            [
-                0..2: Range { bind_left: true, bind_right: true },
-                2..3: Literal(Integer(2)),
-            ],
-        )
-        "###);
-
-        assert_debug_snapshot!(TokenVec(lexer().parse("1..").unwrap()), @r###"
-        TokenVec(
-            [
-                0..1: Literal(Integer(1)),
-                1..3: Range { bind_left: true, bind_right: true },
-            ],
-        )
-        "###);
-
-        assert_debug_snapshot!(TokenVec(lexer().parse("in ..5").unwrap()), @r###"
-        TokenVec(
-            [
-                0..2: Ident("in"),
-                2..5: Range { bind_left: false, bind_right: true },
-                5..6: Literal(Integer(5)),
-            ],
-        )
-        "###);
-    }
-}
