+++ prqlc/prqlc-parser/src/parser/types.rs
@@ -1,9 +1,11 @@
 use chumsky::prelude::*;
-use prqlc_ast::*;
+
+use crate::error::parse_error::PError;
+use crate::lexer::lr::TokenKind;
+use crate::parser::expr::ident;
+use crate::parser::pr::*;
 
 use super::common::*;
-use crate::err::parse_error::PError;
-use crate::expr::ident;
 
 pub fn type_expr() -> impl Parser<TokenKind, Ty, Error = PError> {
     recursive(|nested_type_expr| {
