+++ prqlc/prqlc-parser/src/test.rs
@@ -3,9 +3,8 @@
 use insta::{assert_debug_snapshot, assert_yaml_snapshot};
 use itertools::Itertools;
 
-use crate::ast::expr::*;
-use crate::ast::stmt::*;
-use crate::err::error::*;
+use crate::error::Error;
+use crate::parser::pr::{Expr, FuncCall, Stmt};
 
 /// Helper that does not track source_ids
 fn parse_single(source: &str) -> Result<Vec<Stmt>, Vec<Error>> {
