+++ prqlc/prqlc/Cargo.toml
@@ -30,7 +30,6 @@ cli = [
 ]
 default = ["cli"]
 serde_yaml = [
-  "prqlc-ast/serde_yaml",
   "prqlc-parser/serde_yaml",
   "dep:serde_yaml",
 ]
@@ -40,7 +39,6 @@ test-dbs = ["anyhow", "duckdb", "glob", "rusqlite", "tokio"]
 test-dbs-external = ["anyhow", "duckdb", "glob", "mysql", "pg_bigdecimal", "postgres", "rusqlite", "tiberius", "tokio", "tokio-util"]
 
 [dependencies]
-prqlc-ast = {path = "../prqlc-ast", version = "0.12.3"}
 prqlc-parser = {path = "../prqlc-parser", version = "0.12.3"}
 
 anstream = {version = "0.6.14", features = ["auto"]}
