+++ prqlc/prqlc/src/ir/pl/expr.rs
@@ -1,12 +1,12 @@
 use std::collections::HashMap;
 
 use enum_as_inner::EnumAsInner;
-use prqlc_ast::token::Literal;
+
+use prqlc_parser::parser::generic;
 use serde::{Deserialize, Serialize};
 
 use super::{Lineage, TransformCall};
-use crate::ast::generic;
-use crate::ast::{GenericTypeParam, Ident, Span, Ty};
+use crate::ast::{GenericTypeParam, Ident, Literal, Span, Ty};
 use crate::codegen::write_ty;
 
 // The following code is tested by the tests_misc crate to match expr.rs in prqlc_ast.
