+++ prqlc/prqlc/src/ir/pl/utils.rs
@@ -1,5 +1,5 @@
 use super::{Expr, ExprKind, FuncCall};
-use crate::ast::expr::Ident;
+use crate::ast::Ident;
 
 pub fn maybe_binop(left: Option<Expr>, op_name: &[&str], right: Option<Expr>) -> Option<Expr> {
     match (left, right) {
