+++ prqlc/prqlc/src/ir/rq/expr.rs
@@ -1,9 +1,9 @@
 use enum_as_inner::EnumAsInner;
+use prqlc_parser::lexer::lr::Literal;
 use serde::{Deserialize, Serialize};
 
-use super::super::pl::Literal;
 use super::CId;
-use crate::Span;
+use crate::{ast, Span};
 
 /// Analogous to [crate::ir::pl::Expr], but with fewer kinds.
 #[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
@@ -12,9 +12,9 @@ pub struct Expr {
     pub span: Option<Span>,
 }
 
-pub(super) type Range = crate::ast::generic::Range<Expr>;
-pub(super) type InterpolateItem = crate::ast::generic::InterpolateItem<Expr>;
-pub(super) type SwitchCase = crate::ast::generic::SwitchCase<Expr>;
+pub(super) type Range = ast::generic::Range<Expr>;
+pub(super) type InterpolateItem = ast::generic::InterpolateItem<Expr>;
+pub(super) type SwitchCase = ast::generic::SwitchCase<Expr>;
 
 #[derive(Debug, PartialEq, Clone, Serialize, Deserialize, EnumAsInner)]
 pub enum ExprKind {
