+++ prqlc/prqlc/src/semantic/lowering.rs
@@ -4,22 +4,22 @@ use std::iter::zip;
 
 use enum_as_inner::EnumAsInner;
 use itertools::Itertools;
+use prqlc_parser::lexer::lr::Literal;
 
 use crate::ast::TyTupleField;
 use crate::compiler_version;
 use crate::ir::decl::{self, DeclKind, Module, RootModule, TableExpr};
 use crate::ir::generic::{ColumnSort, WindowFrame};
+use crate::ir::pl::TableExternRef::LocalTable;
 use crate::ir::pl::{self, Ident, Lineage, LineageColumn, PlFold, QueryDef};
 use crate::ir::rq::{
     self, CId, RelationColumn, RelationLiteral, RelationalQuery, TId, TableDecl, Transform,
 };
 use crate::semantic::write_pl;
 use crate::utils::{toposort, IdGenerator};
-use crate::{
-    ast::generic::{InterpolateItem, Range, SwitchCase},
-    ir::pl::TableExternRef::LocalTable,
-};
+
 use crate::{Error, Reason, Result, Span, WithErrorInfo};
+use prqlc_parser::parser::generic::{InterpolateItem, Range, SwitchCase};
 
 /// Convert a resolved expression at path `main_path` relative to `root_mod`
 /// into RQ and make sure that:
@@ -981,7 +981,7 @@ impl Lowerer {
 
 fn str_lit(string: String) -> rq::Expr {
     rq::Expr {
-        kind: rq::ExprKind::Literal(pl::Literal::String(string)),
+        kind: rq::ExprKind::Literal(Literal::String(string)),
         span: None,
     }
 }
