+++ prqlc/prqlc/src/sql/srq/ast.rs
@@ -8,11 +8,11 @@ use itertools::Itertools;
 use serde::Serialize;
 
 use super::context::RIId;
-use crate::ast::generic::InterpolateItem;
 use crate::ir::generic::ColumnSort;
 use crate::ir::pl::JoinSide;
 use crate::ir::rq::{self, fold_column_sorts, RelationLiteral, RqFold};
 use crate::Result;
+use prqlc_parser::parser::generic::InterpolateItem;
 
 #[derive(Debug, Clone, Serialize)]
 pub struct SqlQuery {
