+++ prqlc/prqlc/tests/integration/resolving.rs
@@ -1,5 +1,6 @@
 use insta::assert_snapshot;
 use prqlc::ErrorMessages;
+use prqlc_parser::parser::pr;
 
 // equivalent to prqlc debug resolve
 fn resolve(prql_source: &str) -> Result<String, ErrorMessages> {
@@ -16,7 +17,7 @@ fn resolve(prql_source: &str) -> Result<String, ErrorMessages> {
     prqlc::pl_to_prql(&root_module)
 }
 
-fn drop_module_defs(stmts: &mut Vec<prqlc_parser::ast::stmt::Stmt>, to_drop: &[&str]) {
+fn drop_module_defs(stmts: &mut Vec<pr::Stmt>, to_drop: &[&str]) {
     stmts.retain(|x| {
         x.kind
             .as_module_def()
