+++ prqlc/prqlc/tests/integration/static_eval.rs
@@ -1,6 +1,6 @@
 use insta::assert_yaml_snapshot;
 use prqlc::ir::{constant::ConstExpr, decl::RootModule};
-use prqlc_parser::ast::{Expr, VarDef};
+use prqlc_parser::parser::pr::{Expr, VarDef};
 
 #[track_caller]
 fn static_eval(prql_source: &str) -> ConstExpr {
